+++ .docker/docker-entrypoint.sh
@@ -5,26 +5,33 @@ if [ "$VERBOSE" == "yes" ];then
 fi
 source /usr/local/bin/docker-entrypoint.sh
 # call main bits of the mariadb entrypoint to have DB initialized
-docker_setup_env
-docker_create_db_directories
+docker_setup_env "mariadbd"
+docker_create_db_directories "mariadbd"
 # there's no database, so it needs to be initialized
 if [ -z "$DATABASE_ALREADY_EXISTS" ]; then
-        docker_verify_minimum_env
-
-        docker_mariadb_init "$@"
+    docker_verify_minimum_env "mariadbd"
+    docker_mariadb_init "mariadbd"
 elif _check_if_upgrade_is_needed; then
-        docker_mariadb_upgrade "$@"
+    docker_mariadb_upgrade "mariadbd"
 fi
-mariadbd &
-until healthcheck.sh --connect --innodb_initialized;
+mariadbd --user=${MARIADB_USER} --datadir=${MYSQL_DATA_DIR} --socket=/tmp/mysql.sock --console &
+until mariadb-admin ping --user="${MARIADB_USER}" --password="${MARIADB_PASSWORD}" --socket="/tmp/mysql.sock" --silent;
 do
     echo "waiting for mariadb"
+    echo "${MARIADB_USER} ${MARIADB_PASSWORD}"
     sleep 1
 done
 
 solr start -s ${SOLR_HOME}
 if [ "${IS_BINDER}" = "true" ];then
-    python $EVAL_HOME/ingest_dummy_data.py /mnt/data4freva
+    # wait for solr to be up
+    solr_status_url="http://localhost:8983/solr/admin/cores?action=STATUS"
+    until curl -s "$solr_status_url" > /dev/null; do
+        echo "Waiting for Solr to be available..."
+        sleep 2
+    done
+    echo "Solr is up!"
+    python $EVAL_HOME/ingest_dummy_data.py /mnt/data4frev
     python $EVAL_HOME/dummy_user_data.py
     for i in 1 2 3 4 5; do
         freva plugin dummyplugin the_number=$i