+++ src/evaluation_system/tests/plugin_command_test.py
@@ -31,22 +31,9 @@ def test_cli(dummy_plugin, capsys, dummy_config, caplog):
         assert "the_number" in output
         assert "13" in output
         assert os.getpid() == 12345
-        out_path = (
-            Path(dummy_config.get("scheduler_output_dir"))
-            / "dummyplugin"
-            / "DummyPlugin-12345.local"
-        )
-        try:
-            out_path.unlink()
-        except FileNotFoundError:
-            pass
         plugin_cli(["dummyplugin", "the_number=13"])
-        interactive = capsys.readouterr().out
-        assert dummy_plugin.plugin_output_file == out_path
-        assert out_path.exists()
-        with out_path.open() as f:
-            assert interactive == f.read()
-        out_path.unlink()
+        out_path = dummy_plugin.plugin_output_file
+        assert str(out_path).endswith("-12345.local")
         plugin_cli(["dummyplugin", "the_number=13", "--batchmode"])
         output = capsys.readouterr().out
         assert "tail -f" in output
@@ -132,7 +119,7 @@ def test_plugin_status(dummy_env, caplog) -> None:
 
     import freva
 
-    res = freva.run_plugin("dummyplugin", the_number=2, other=-5, batchmode=True)
+    res = freva.run_plugin("dummyplugin", the_number=2, other=-15, batchmode=True)
     with pytest.raises(ValueError):
         res.wait(2)
     assert res.status == "running"
@@ -176,6 +163,7 @@ def test_empty_status(dummy_history, capsys) -> None:
     assert res.stdout == ""
     assert res.batch_id is None
     assert res.job_script == ""
+    assert isinstance(res.history_id, int)
     res.kill()
     assert res.plugin == ""
     assert res.version == (0, 0, 0)